# src/openapi.yaml
openapi: 3.0.3
info:
  title: Todo MERN API
  description: |
    A production-ready Todo API built with Express.js, MongoDB, and TypeScript.

    ## Authentication
    This API uses JWT-based authentication with access and refresh tokens.
    - **Access Token**: Short-lived (15 minutes), sent in Authorization header
    - **Refresh Token**: Long-lived (30 days), stored in httpOnly cookie

    ## Getting Started
    1. Register a new account using `/api/auth/register`
    2. Use the returned access token in the Authorization header: `Bearer <token>`
    3. When the access token expires, use `/api/auth/refresh` to get a new one
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Uses the refresh token from cookie to generate new tokens
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your access token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            body:
              - 'email: Invalid email format'
              - 'password: Password must be at least 8 characters'
        stack:
          type: string
          description: Stack trace (only in development mode)